Protect function

+++ uninstall.php
@@ -8,9 +8,12 @@
     // If preserve settings is false
     if ( false == get_option( 'b3_preserve_settings' ) ) {
         
-        $meta_keys = b3_get_all_custom_meta_keys();
-        foreach ( $meta_keys as $key ) {
-            delete_option( $key );
+        $meta_keys = [];
+        if ( function_exists( 'b3_get_all_custom_meta_keys' ) ) {
+            $meta_keys = b3_get_all_custom_meta_keys();
+            foreach ( $meta_keys as $key ) {
+                delete_option( $key );
+            }
         }
         
         $roles = array(
